pyproject.toml
[tool.poetry]
name = "comp4"
version = "0.1.0"
description = "Multi-Factor Compensation Tool for UK equity & fixed-income portfolio managers"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.111"}
sqlalchemy = {version = "^2.0", extras = ["asyncio"]}
asyncpg = "^0.29"
pydantic = "^2.7"
alembic = "^1.13"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.9.0"
isort = "^5.12.0"
ruff = "^0.0.291"
mypy = "^1.5.0"
pytest-cov = "^4.1.0"
aiosqlite = "^0.21.0"
pytest-asyncio = ">=0.21.0"

[tool.poetry.scripts]
comp4-api = "app.main:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 88

[tool.ruff]
line-length = 88
target-version = "py311"
select = ["E", "F", "B", "I"]

[tool.mypy]
python_version = "3.11"
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
strict_optional = true
warn_redundant_casts = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
addopts = "--cov=app --cov-report=term-missing"
asyncio_mode = "auto"
